lexicon added - it is from the source in the description
edited and parse - and still being worked on... seeing if there is an easier import option, but this was the biggest free source I've found - it is from this website:
https://conwaylife.com/ref/lexicon/lex_home.htm
but really wish they had labeled the <pre> tags - would have been so much easier.
i'm still filtering out the page so the simple art parser works correctly.